+++ 05-examples/demos/02-bpmn/src/main/scala/camundala/examples/demos/bpmn/DateExample.scala
@@ -1,0 +1,25 @@
+package camundala.examples.demos.bpmn
+
+import camundala.bpmn.*
+import camundala.domain.*
+import java.time.LocalDateTime
+
+//TODO this test throws a java.net.ConnectException
+object DateExample extends BpmnDsl:
+
+  case class Input(inDate: LocalDateTime = LocalDateTime.parse("2013-12-12T12:12:12"))
+
+  object Input:
+    given ApiSchema[Input] = deriveSchema
+    given InOutCodec[Input] = deriveCodec
+
+  lazy val DateExampleDMN = singleEntry(
+    decisionDefinitionKey = "DateExample",
+    in = Input(),
+    out = "2012-12-12T12:12:12.000+0100" //Camunda returns String - LocalDateTime.parse("2012-12-12T12:12:12")
+  )
+
+
+
+
+end DateExample