+++ 05-examples/demos/03-simulation/src/test/scala/camundala/examples/demos/simulation/BasicAuthExampleSimulation.scala
@@ -1,0 +1,16 @@
+package camundala.examples.demos.simulation
+
+import camundala.examples.demos.bpmn.DateExample.DateExampleDMN
+import camundala.simulation.*
+import camundala.simulation.custom.*
+
+// exampleDemosSimulation/test
+// exampleDemosSimulation/testOnly *BasicAuthExampleSimulation
+class BasicAuthExampleSimulation extends BasicSimulationDsl:
+
+  simulate (
+    DateExampleDMN
+  )
+
+  override implicit def config =
+    super.config.withPort(8033)