+++ 05-examples/demos/src/it/scala/camundala/examples/demos/DemosSimulation.scala
@@ -1,7 +1,0 @@
-package camundala.examples.demos
-
-import camundala.simulation.custom.CustomSimulation
-
-trait DemosSimulation extends CustomSimulation :
-  override implicit def config =
-    super.config.withPort(8887)