+++ 05-examples/invoice/02-bpmn/src/main/scala/camundala/examples/invoice/bpmn/exports.scala
@@ -1,0 +1,11 @@
+package camundala.examples.invoice.bpmn
+
+import camundala.domain.*
+
+enum InvoiceCategory:
+  case `Travel Expenses`, Misc, `Software License Costs`
+
+object InvoiceCategory:
+  given ApiSchema[InvoiceCategory] = deriveEnumSchema
+  given InOutCodec[InvoiceCategory] = deriveEnumCodec
+end InvoiceCategory