+++ 05-examples/invoice/03-worker/src/main/scala/camundala/examples/invoice/worker/InvoiceReceiptWorker.scala
@@ -1,0 +1,13 @@
+package camundala.examples.invoice.worker
+
+import camundala.bpmn
+import camundala.examples.invoice.bpmn.InvoiceReceipt.*
+import camundala.worker.InitWorkerDsl
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class InvoiceReceiptWorker extends InvoiceWorkerHandler, InitWorkerDsl[In, Out]:
+
+  lazy val inOut = example
+
+end InvoiceReceiptWorker