+++ 05-examples/invoice/03-worker/src/main/scala/camundala/examples/invoice/worker/StarWarsPeopleDetailWorker.scala
@@ -1,0 +1,39 @@
+package camundala.examples.invoice.worker
+
+import camundala.domain.*
+import camundala.examples.invoice.bpmn.StarWarsPeopleDetail.*
+import camundala.worker.*
+import camundala.worker.CamundalaWorkerError.*
+import org.springframework.context.annotation.Configuration
+import sttp.client3.UriContext
+import sttp.model.Uri
+
+@Configuration
+class StarWarsPeopleDetailWorker extends InvoiceWorkerHandler,
+      ServiceWorkerDsl[In, Out, NoInput, ServiceOut]:
+
+  lazy val serviceTask = example
+
+  def apiUri(in: In): Uri = uri"https://swapi.dev/api/people/${in.id}"
+
+  override protected def querySegments: Seq[QuerySegmentOrParam] =
+    QuerySegmentOrParam.keys("id") ++
+      QuerySegmentOrParam.keyValues("a" -> 1, "b" -> true) ++
+      QuerySegmentOrParam.values(12, false, null)
+    
+  override def validate(in: In): Either[ValidatorError, In] =
+    if in.id <= 0 then
+      Left(ValidatorError("The search id for People must be > 0!"))
+    else
+      super.validate(in)
+
+  override def outputMapper(
+      serviceOut: ServiceResponse[ServiceOut],
+      in: In
+  ): Either[ServiceMappingError, Out] =
+    Right(Out.Success(serviceOut.outputBody, serviceOut.headers.getOrElse("fromHeader", "---")))
+
+  override protected def inputHeaders(in: In): Map[String, String] =
+    Map("test-db-id" -> in.id.toString)
+
+end StarWarsPeopleDetailWorker