+++ 05-examples/invoice/04-c7-spring/src/main/scala/camundala/examples/invoice/c7/Invoice7ExampleApplication.scala
@@ -1,0 +1,23 @@
+package camundala.examples.invoice.c7
+
+import org.camunda.bpm.example.invoice.c7.InvoiceProcessApplication
+import org.camunda.bpm.spring.boot.starter.annotation.EnableProcessApplication
+import org.camunda.bpm.spring.boot.starter.event.PostDeployEvent
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.context.event.EventListener
+
+@SpringBootApplication
+@EnableProcessApplication
+class Invoice7ExampleApplication
+
+object Invoice7ExampleApplication:
+
+  def main(args: Array[String]): Unit =
+    SpringApplication.run(classOf[Invoice7ExampleApplication], args: _*)
+
+  lazy val invoicePa = new InvoiceProcessApplication()
+
+  @EventListener
+  def onPostDeploy(event: PostDeployEvent): Unit =
+    invoicePa.startFirstProcess(event.getProcessEngine)