+++ 05-examples/twitter/03-api/src/main/scala/camundala/examples/twitter/api/ProjectApiCreator.scala
@@ -1,0 +1,35 @@
+package camundala.examples.twitter.api
+
+import camundala.domain.*
+import camundala.bpmn.*
+import camundala.api.*
+import camundala.examples.twitter.bpmn.*
+import camundala.examples.twitter.bpmn.TwitterApi.*
+
+
+// exampleTwitterApi/run
+object ProjectApiCreator extends DefaultApiCreator:
+
+  val projectName = "twitter-example-c7"
+
+  val title = "Twitter Process API C7"
+
+  val version = "1.0"
+
+  document(
+    api(twitterDemoProcess)(
+      ReviewTweetApprovedUT
+    )
+  )
+
+  override lazy val apiConfig: ApiConfig =
+    super.apiConfig
+      .withBasePath(os.pwd / "05-examples" / "twitter")
+      .withPort(8887)
+
+  private lazy val `Tweet accepted` = reviewTweetApprovedUT
+  private lazy val ReviewTweetApprovedUT =
+    `Tweet accepted`
+      .withOutExample("Tweet rejected", ReviewedTweet(approved = false))
+
+