+++ 05-examples/twitter/04-c8-spring/src/main/java/camundala/examples/twitter/business/DuplicateTweetException.java
@@ -1,0 +1,11 @@
+package camundala.examples.twitter.business;
+
+public class DuplicateTweetException extends Exception {
+    public DuplicateTweetException(String message) {
+        super(message);
+    }
+
+    public DuplicateTweetException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}