+++ 05-examples/twitter/04-c8-spring/src/main/java/camundala/examples/twitter/business/TwitterService.java
@@ -1,0 +1,11 @@
+package camundala.examples.twitter.business;
+
+
+/**
+ * Publish content on Twitter.
+ */
+public interface TwitterService {
+
+  void tweet(String content) throws DuplicateTweetException;
+
+}